Implemented separate input management for ball and mouse inputs
The Hit() method in I_SmartWallInteractable now has a new parameter: InputType, which hold the type of input that was being made to hit the object.
There are 4 InputTypes:
- Ball, hopefully obvious.
- MouseDown, the first frame a mouse input is detected.
- MouseHold, returns true for as long as the mouse is being held down.
- MouseUp, the first frame a mouse input is let go.

NOTE: holding down the mouse button and moving the mouse outside of the object detection range does NOT cause the MouseUp state to trigger.